backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.1
openai==1.3.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.0
requests==2.31.0
gunicorn==21.2.0
python-multipart==0.0.6
